+++ docs/release-notes.md
@@ -6,6 +6,7 @@
 
 ### For developers:
 * The `RequiredApiVersion` `nickel.json` field is no longer used on non-assembly mods.
+* The `RequiredApiVersion` `nickel.json` field is now optional on assembly mods.
 * Improved sprite asset registration and error handling.
 
 ## 1.6.1